SERVICES-2668: move repository service and update imports
