+++ src/modules/analytics-indexer/analytics.indexer.module.ts
@@ -13,6 +13,8 @@ import { IndexerPriceDiscoveryService } from './services/indexer.price.discovery
 import { IndexerSwapHandlerService } from './services/event-handlers/indexer.swap.handler.service';
 import { IndexerLiquidityHandlerService } from './services/event-handlers/indexer.liquidity.handler.service';
 import { IndexerPriceDiscoveryHandlerService } from './services/event-handlers/indexer.price.discovery.handler.service';
+import { IndexerSessionRepositoryService } from './services/indexer.session.repository.service';
+import { IndexerPersistenceService } from './services/indexer.persistence.service';
 import { AnalyticsIndexerController } from './analytics.indexer.controller';
 import { DatabaseModule } from 'src/services/database/database.module';
 import { MongooseModule } from '@nestjs/mongoose';
@@ -20,10 +22,11 @@ import {
     IndexerSession,
     IndexerSessionSchema,
 } from './schemas/indexer.session.schema';
-import { IndexerSessionRepositoryService } from 'src/services/database/repositories/indexer.session.repository';
+import { MXCommunicationModule } from 'src/services/multiversx-communication/mx.communication.module';
 
 @Module({
     imports: [
+        MXCommunicationModule,
         PairModule,
         RouterModule,
         TokenModule,
@@ -45,6 +48,7 @@ import { IndexerSessionRepositoryService } from 'src/services/database/repositor
         IndexerLiquidityHandlerService,
         IndexerPriceDiscoveryHandlerService,
         IndexerSessionRepositoryService,
+        IndexerPersistenceService,
     ],
     exports: [IndexerService],
     controllers: [AnalyticsIndexerController],
